Fixed some more warnings GCC was giving about unused variables and signed/unsigned in the M4A handlers

+++ libAudio/loadM4A.cpp
@@ -184,7 +184,7 @@ int MP4DecRead(void *MP4File, void *DataOut, int64_t DataOutLen, int64_t *Read, 
  */
 int MP4DecWrite(void *MP4File, const void *DataIn, int64_t DataInLen, int64_t *Written, int64_t)
 {
-	if (fwrite(DataIn, 1, (size_t)DataInLen, (FILE *)MP4File) != DataInLen)
+	if (fwrite(DataIn, 1, (size_t)DataInLen, (FILE *)MP4File) != (size_t)DataInLen)
 		return TRUE;
 	*Written = DataInLen;
 	return FALSE;
